refactor(adapters): delete shims + migrate consumers to lyra.outbound (T27, refs #1279)
Deletes shims from S1:
- src/lyra/adapters/shared/_shared_streaming_emitter.py
- src/lyra/adapters/shared/_shared_streaming.py

Migrates all consumers to import from lyra.outbound.{emitter,throttle}.
Consumer set: adapters.shared.__init__/_base_outbound/_shared, telegram.py,
discord/adapter.py, telegram_outbound.py, discord_outbound.py, plus 8 tests.

_shared_streaming_state.py keeps IntermediateTextState + StreamState (per
spec Open Q 2). STREAMING_EDIT_INTERVAL re-export shim removed; emitter.py
imports it from lyra.outbound.throttle at module bottom. classify_stream_error
legacy wrapper removed (deferred import inside _shared_streaming_state.py
state-machine consumer).

553 tests pass.
