+++ tests/adapters/test_streaming_session.py
@@ -17,13 +17,14 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 
 from lyra.adapters.nats.nats_stream_decoder import decode_stream_events
-from lyra.adapters.shared._shared_streaming import PlatformCallbacks, StreamingSession
 from lyra.core.exceptions import StreamChunkTimeout
 from lyra.core.messaging.message import GENERIC_ERROR_REPLY, OutboundMessage
 from lyra.core.messaging.render_events import (
     TextDeltaRenderEvent,
     TextEndRenderEvent,
 )
+from lyra.outbound.emitter import OutboundEmitter as StreamingSession
+from lyra.outbound.emitter import PlatformCallbacks
 
 # ---------------------------------------------------------------------------
 # Helpers
